clean abi and fix signers enriched to ensure it has proper data
