+++ app/lib/farcaster/abi/signed-key-request-abi.ts
@@ -1,0 +1,12 @@
+export const signedKeyRequestAbi = [
+  {
+    components: [
+      { name: "requestFid", type: "uint256" },
+      { name: "requestSigner", type: "address" },
+      { name: "signature", type: "bytes" },
+      { name: "deadline", type: "uint256" },
+    ],
+    name: "SignedKeyRequest",
+    type: "tuple",
+  },
+] as const;
